macosx: fix 32-bit legacy compile issues

+++ macosx/Badger.mm
@@ -49,7 +49,7 @@
     NSParameterAssert(torrent != nil);
 
     [self.fHashes addObject:torrent.hashString];
-    [[NSApp dockTile] setBadgeLabel:[NSString localizedStringWithFormat:@"%lu", self.fHashes.count]];
+    [[NSApp dockTile] setBadgeLabel:[NSString localizedStringWithFormat:@"%lu", static_cast<unsigned long>(self.fHashes.count)]];
 }
 
 - (void)removeTorrent:(Torrent*)torrent
@@ -59,7 +59,7 @@
         [self.fHashes removeObject:torrent.hashString];
         if (self.fHashes.count > 0)
         {
-            [[NSApp dockTile] setBadgeLabel:[NSString localizedStringWithFormat:@"%lu", self.fHashes.count]];
+            [[NSApp dockTile] setBadgeLabel:[NSString localizedStringWithFormat:@"%lu", static_cast<unsigned long>(self.fHashes.count)]];
         }
         else
         {
